+++ synology_cli/parameters/webservice.py
@@ -1,4 +1,6 @@
 
+ENTRY_URL = '{url}/webapi/entry.cgi'
+
 LOGIN_PARAMS = {
     'api': 'SYNO.API.Auth',
     'version': 7,
